+++ gluecodium/src/test/java/com/here/gluecodium/validator/LimeLambdaValidatorDocsTest.kt
@@ -19,6 +19,7 @@
 
 package com.here.gluecodium.validator
 
+import com.here.gluecodium.common.LimeLogger
 import com.here.gluecodium.generator.common.GeneratorOptions
 import com.here.gluecodium.model.lime.LimeAttributeType
 import com.here.gluecodium.model.lime.LimeAttributes
@@ -31,7 +32,9 @@ import com.here.gluecodium.model.lime.LimeModel
 import com.here.gluecodium.model.lime.LimePath
 import com.here.gluecodium.model.lime.LimeReturnType
 import com.here.gluecodium.model.lime.LimeStruct
+import io.mockk.justRun
 import io.mockk.mockk
+import io.mockk.verify
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
@@ -54,7 +57,14 @@ class LimeLambdaValidatorDocsTest {
                 path = lambdaPath,
                 comment = limeComment,
                 returnType = LimeReturnType(typeRef = LimeBasicTypeRef.INT, comment = LimeComment("Important integer")),
-                parameters = listOf(LimeLambdaParameter(path = lambdaPath.child("param1"), typeRef = LimeBasicTypeRef.FLOAT)),
+                parameters =
+                    listOf(
+                        LimeLambdaParameter(
+                            path = lambdaPath.child("param1"),
+                            isNamedParameter = true,
+                            typeRef = LimeBasicTypeRef.FLOAT,
+                        ),
+                    ),
             )
         allElements[lambdaPath.toString()] = limeLambda
 
@@ -75,7 +85,14 @@ class LimeLambdaValidatorDocsTest {
                 path = lambdaPath,
                 comment = limeComment,
                 returnType = LimeReturnType(typeRef = LimeBasicTypeRef.INT, comment = LimeComment("Important integer")),
-                parameters = listOf(LimeLambdaParameter(path = lambdaPath.child("param1"), typeRef = LimeBasicTypeRef.FLOAT)),
+                parameters =
+                    listOf(
+                        LimeLambdaParameter(
+                            path = lambdaPath.child("param1"),
+                            isNamedParameter = true,
+                            typeRef = LimeBasicTypeRef.FLOAT,
+                        ),
+                    ),
             )
         allElements[lambdaPath.toString()] = limeLambda
 
@@ -164,11 +181,13 @@ class LimeLambdaValidatorDocsTest {
                     listOf(
                         LimeLambdaParameter(
                             path = lambdaPath.child("param1"),
+                            isNamedParameter = true,
                             typeRef = LimeBasicTypeRef.FLOAT,
                             comment = LimeComment("Some param"),
                         ),
                         LimeLambdaParameter(
                             path = lambdaPath.child("param2"),
+                            isNamedParameter = true,
                             typeRef = LimeBasicTypeRef.FLOAT,
                             comment = LimeComment("Another param"),
                         ),
@@ -196,7 +215,14 @@ class LimeLambdaValidatorDocsTest {
                 comment = limeComment,
                 attributes = attributes,
                 returnType = LimeReturnType(typeRef = LimeBasicTypeRef.INT, comment = LimeComment("Important integer")),
-                parameters = listOf(LimeLambdaParameter(path = lambdaPath.child("param1"), typeRef = LimeBasicTypeRef.FLOAT)),
+                parameters =
+                    listOf(
+                        LimeLambdaParameter(
+                            path = lambdaPath.child("param1"),
+                            isNamedParameter = true,
+                            typeRef = LimeBasicTypeRef.FLOAT,
+                        ),
+                    ),
             )
         allElements[lambdaPath.toString()] = limeLambda
 
@@ -219,7 +245,14 @@ class LimeLambdaValidatorDocsTest {
                 path = lambdaPath,
                 comment = limeComment,
                 returnType = LimeReturnType(typeRef = LimeBasicTypeRef.INT, comment = LimeComment("Important integer")),
-                parameters = listOf(LimeLambdaParameter(path = lambdaPath.child("param1"), typeRef = LimeBasicTypeRef.FLOAT)),
+                parameters =
+                    listOf(
+                        LimeLambdaParameter(
+                            path = lambdaPath.child("param1"),
+                            isNamedParameter = true,
+                            typeRef = LimeBasicTypeRef.FLOAT,
+                        ),
+                    ),
             )
 
         val attributes = LimeAttributes.Builder().addAttribute(LimeAttributeType.INTERNAL).build()
@@ -248,7 +281,14 @@ class LimeLambdaValidatorDocsTest {
                 path = lambdaPath,
                 comment = limeComment,
                 returnType = LimeReturnType(typeRef = LimeBasicTypeRef.INT, comment = LimeComment("Important integer")),
-                parameters = listOf(LimeLambdaParameter(path = lambdaPath.child("param1"), typeRef = LimeBasicTypeRef.FLOAT)),
+                parameters =
+                    listOf(
+                        LimeLambdaParameter(
+                            path = lambdaPath.child("param1"),
+                            isNamedParameter = true,
+                            typeRef = LimeBasicTypeRef.FLOAT,
+                        ),
+                    ),
             )
 
         val nestedStruct = LimeStruct(path = nestedStructPath, lambdas = listOf(limeLambda))
@@ -265,6 +305,90 @@ class LimeLambdaValidatorDocsTest {
         val result = validator.validate(limeModel)
 
         // Then docs validation passes (for lambdas from types nested in internal ones it is skipped).
+        assertTrue(result)
+    }
+
+    @Test
+    fun validateDocumentationCommentForUnnamedParameterWhenWerrorIsEnabled() {
+        // Given LimeLambda with parameter with default name that is documented.
+        val lambdaPath = LimePath(listOf(), listOf("SomeLambda"))
+        val parameter =
+            LimeLambdaParameter(
+                path = lambdaPath.child("p0"),
+                comment = LimeComment("Some unnamed param"),
+                isNamedParameter = false,
+                typeRef = LimeBasicTypeRef.FLOAT,
+            )
+
+        val limeLambda =
+            LimeLambda(
+                path = lambdaPath,
+                comment = limeComment,
+                returnType = LimeReturnType(typeRef = LimeBasicTypeRef.INT, comment = LimeComment("Important integer")),
+                parameters = listOf(parameter),
+            )
+        allElements[lambdaPath.toString()] = limeLambda
+
+        // When validating it with werror flag enabled.
+        val logger: LimeLogger = mockk()
+        justRun { logger.error(limeLambda, any()) }
+
+        val validator = LimeLambdaValidator(logger = logger, generatorOptions = generatorOptions)
+        val result = validator.validate(limeModel)
+
+        // Then validation fails.
+        verify(exactly = 1) {
+            logger.error(
+                limeLambda,
+                match {
+                    it.startsWith(
+                        "Default parameter name '${parameter.name}' is documented. Please set an explicit name for documented parameters;",
+                    )
+                },
+            )
+        }
+        assertFalse(result)
+    }
+
+    @Test
+    fun validateDocumentationCommentForUnnamedParameterWhenWerrorIsDisabled() {
+        // Given LimeLambda with parameter with default name that is documented.
+        val lambdaPath = LimePath(listOf(), listOf("SomeLambda"))
+        val parameter =
+            LimeLambdaParameter(
+                path = lambdaPath.child("p0"),
+                comment = LimeComment("Some unnamed param"),
+                isNamedParameter = false,
+                typeRef = LimeBasicTypeRef.FLOAT,
+            )
+
+        val limeLambda =
+            LimeLambda(
+                path = lambdaPath,
+                comment = limeComment,
+                returnType = LimeReturnType(typeRef = LimeBasicTypeRef.INT, comment = LimeComment("Important integer")),
+                parameters = listOf(parameter),
+            )
+        allElements[lambdaPath.toString()] = limeLambda
+
+        // When validating it with werror flag disabled.
+        val logger: LimeLogger = mockk()
+        justRun { logger.warning(limeLambda, any()) }
+
+        val validator = LimeLambdaValidator(logger = logger)
+        val result = validator.validate(limeModel)
+
+        // Then validation succeeds with warning.
+        verify(exactly = 1) {
+            logger.warning(
+                limeLambda,
+                match {
+                    it.startsWith(
+                        "Default parameter name '${parameter.name}' is documented. Please set an explicit name for documented parameters;",
+                    )
+                },
+            )
+        }
         assertTrue(result)
     }
 }